requirements.txt
chromadb==0.6.3
ollama==0.4.7
langchain_community==0.3.18
langchain_text_splitters==0.3.6
pypdf==5.3.0
Django==5.1.6
djangorestframework==3.15.2
langchain-community==0.3.18
langchain-text-splitters==0.3.6
dotenv==0.9.9
python-dotenv==1.0.1